% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adaptMCMC}
\alias{adaptMCMC}
\title{Metropolis-Hastings MCMC with Adaptive Proposal Kernel}
\usage{
adaptMCMC(
  target,
  init_theta,
  covmat,
  n_iterations,
  adapt_size_start,
  acceptance_rate_weight,
  acceptance_window,
  info,
  seedMH = 1,
  eps = 0.05
)
}
\arguments{
\item{target}{the a R function that must return the log likelihood of the target distribution we want to sample from}

\item{init_theta}{a vector of initial parameter values}

\item{covmat}{the covariance matrix of the proposal distribution}

\item{n_iterations}{the integer value, how long to run the chain}

\item{adapt_size_start}{the number of accepted jumps after which to begin adapting size of proposal covariance matrix}

\item{info}{the print information on chain every X n_iterations}

\item{seedMH}{the seed of RNG}

\item{eps}{the a constant that we may choose very small compared to the size of S}
}
\value{
R list: theta_trace,sigma_empirical,acceptance_rate
}
\description{
Metropolis-Hastings MCMC with Adaptive Proposal Kernel
}
\examples{
library(mcmcSampler)
p.log <- function(x) {
B <- 0.03 # controls 'bananacity'
-x[1]^2/200 - 1/2*(x[2]+B*x[1]^2-100*B)^2
}
set.seed(123)
 banana1 <- adaptMCMC(target=p.log,init_theta=c(10,10),covmat=diag(c(1,1)),n_iterations=1e3,adapt_size_start=10,acceptance_rate_weight=0,acceptance_window=0,info=1e2, seedMH = 123)

par(mfrow=c(1,2))
x1 <- seq(-15, 15, length=100)
x2 <- seq(-15, 15, length=100)
d.banana <- matrix(apply(expand.grid(x1, x2), 1, p.log), nrow=100)
image(x1, x2, exp(d.banana), col=cm.colors(60))
contour(x1, x2, exp(d.banana), add=TRUE, col=gray(0.6))
lines(banana1$theta_trace, type='l')
matplot(banana1$theta_trace,type="l")
}
