% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtmvnorm_gibbs}
\alias{rtmvnorm_gibbs}
\title{Get a sample of size n from the specified truncated multivariate Gaussian distribution}
\usage{
rtmvnorm_gibbs(n, mu, sigma, lower, upper, init_state)
}
\arguments{
\item{n}{integer number of samples to take}

\item{mu}{mean of distribution}

\item{sigma}{standard deviation of distribution}

\item{lower}{lower truncation bound}

\item{upper}{upper truncation bound}

\item{init_state}{initial starting state of the MCMC chain}
}
\value{
a sample of size n from the specified truncated multivariate Gaussian distribution
}
\description{
Get a sample of size n from the specified truncated multivariate Gaussian distribution
}
\examples{
library(mcmcSampler)
set.seed(100)
mean <- c(-1, 1)
sigma <- matrix(data = c(1.0, -0.5, -0.5, 1.0), ncol = 2, nrow = 2)
rtmvnorm_gibbs(1e4, mean, sigma, c(-Inf,0), c(10,10), c(2,2) )
}
