% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtnorm_gibbs}
\alias{rtnorm_gibbs}
\title{Get a sample of size n from the specificed truncated Gaussian distribution.}
\usage{
rtnorm_gibbs(n, mu, sigma, a, b)
}
\arguments{
\item{n}{integer number of samples to take}

\item{mu}{mean of distribution}

\item{sigma}{standard deviation of distribution}

\item{a}{lower truncation bound}

\item{b}{upper truncation bound}
}
\value{
a sample of size n from the specificed truncated Gaussian distribution
}
\description{
Get a sample of size n from the specificed truncated Gaussian distribution.
}
\examples{
library(mcmcSampler)
set.seed(100)
rtnorm_gibbs(n=100,mu=0, sigma=1, a=2, b=4)
}
