% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMCMC}
\alias{rwMCMC}
\title{Random Walk Metropolis-Hastings MCMC}
\usage{
rwMCMC(target, init_theta, covmat, n_iterations)
}
\arguments{
\item{target}{the R function that must return the log likelihood of the target distribution we want to sample from}

\item{init_theta}{the vector of initial parameter values}

\item{covmat}{the covariance matrix of the proposal distribution}

\item{n_iterations}{the integer value, how long to run the chain}
}
\value{
R list: theta_trace, acceptance_rate
}
\description{
Random Walk Metropolis-Hastings MCMC
}
\examples{
library(mcmcSampler)
p.log <- function(x) {
  B <- 0.03 # controls 'bananacity'
  -x[1]^2/200 - 1/2*(x[2]+B*x[1]^2-100*B)^2
}
sample <- rwMCMC(target=p.log, init_theta=c(10,10), covmat=diag(c(1,1)),n_iterations=1e4)
}
